(#1057) Move ContractIdentifiers
